+++ bartext/bartext.go
@@ -1,3 +1,14 @@
+// Copyright 2015 SeukWon Kang (kasworld@gmail.com)
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//    http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package bartext
 
 import (
@@ -15,7 +26,7 @@ type TextControl struct {
 	bg       htmlcolors.RGBA
 	fg       htmlcolors.RGBA
 	bar      float64
-	aligndir int
+	aligndir uint8
 }
 
 func New(x, y, z int, wx, wy int, s string, font *sdlgui.Font) *TextControl {
@@ -115,7 +126,7 @@ func (tc *TextControl) SetBarText(bar float64, format string, a ...interface{}) 
 	tc.SetBar(bar)
 	tc.SetText(format, a...)
 }
-func (tc *TextControl) SetAlign(align int) {
+func (tc *TextControl) SetAlign(align uint8) {
 	if align == tc.aligndir {
 		return
 	}
